Вынесены тесты на валидацию в gateway
