+++ server/src/test/java/ru/practicum/shareit/request/ItemRequestControllerTest.java
@@ -81,46 +81,6 @@ class ItemRequestControllerTest {
     }
 
     @Test
-    void createItemRequest_whenEmptyItemRequest_thenReturnError() throws Exception {
-        ItemRequestDtoRequest invalidRequest = ItemRequestDtoRequest.builder()
-                .description("")
-                .build();
-
-        mvc.perform(post("/requests")
-                        .content(mapper.writeValueAsString(invalidRequest))
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)
-                        .header(Constant.OWNER_ID, "1"))
-                .andDo(print())
-                .andExpect(status().isBadRequest());
-        verifyNoInteractions(itemRequestService);
-    }
-
-    @Test
-    void createItemRequest_whenBigInvalidItemRequest_thenReturnError() throws Exception {
-        ItemRequestDtoRequest invalidRequest = ItemRequestDtoRequest.builder()
-                .description("Реализовать юнит-тесты для всего кода, содержащего логику. " +
-                        "Выберите те классы, которые содержат в себе нетривиальные методы, условия и ветвления. " +
-                        "В основном это будут классы сервисов. Напишите юнит-тесты на все такие методы, используя моки при необходимости.\n" +
-                        "Реализовать интеграционные тесты, проверяющие взаимодействие с базой данных. Как вы помните, " +
-                        "интеграционные тесты представляют собой более высокий уровень тестирования: их обычно требуется " +
-                        "меньше, но покрытие каждого — больше. Мы предлагаем вам создать по одному интеграционному тесту" +
-                        "для каждого крупного метода в ваших сервисах. Например, для метода getUserItems в классе ItemServiceImpl.")
-                .build();
-
-        mvc.perform(post("/requests")
-                        .content(mapper.writeValueAsString(invalidRequest))
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)
-                        .header(Constant.OWNER_ID, "1"))
-                .andDo(print())
-                .andExpect(status().isBadRequest());
-        verifyNoInteractions(itemRequestService);
-    }
-
-    @Test
     void getRequestsById_whenValidId_thenReturnItemRequest() throws Exception {
         long id = 1L;
